Keep imported CSS merging in one rendering helper
html.ts still carried the imported-CSS merge pipeline while html.test.ts only
reached it through a private method. Moving that logic to a dedicated helper
shrinks the HTML generator and lets the tests exercise the real merge behavior
directly without reaching through private internals.

Constraint: Keep the slice internal to rendering/orchestrator and avoid changing merged CSS output semantics
Rejected: Fold additional HTMLGenerator helper extraction into the same PR | broadens scope beyond one smell and weakens rollback clarity
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Route future imported CSS merge changes through html-imported-css.ts instead of regrowing html.ts
Tested: Targeted html/version tests, fmt, lint, typecheck, verify:quick
Not-tested: Downstream consumer repos because this slice is internal to veryfront-code
