+++ deno.json
@@ -1,6 +1,6 @@
 {
   "name": "veryfront",
-  "version": "0.1.173",
+  "version": "0.1.174",
   "license": "Apache-2.0",
   "nodeModulesDir": "auto",
   "exclude": [
